UC-12 refactor: /entities/Posts via FSD methodology
