+++ src/entities/Posts/model/types.ts
