+++ src/entities/Posts/ui/CircleBlock.tsx
