+++ src/entities/Posts/ui/DescriptionBlock.tsx
