+++ src/entities/Posts/ui/SliderArrow.tsx
