+++ src/entities/Posts/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './CircleBlock'
+export * from './DescriptionBlock'
